remove test suite views
